+++ user_warehouse_restriction/models/res_users.py
@@ -46,9 +46,11 @@ class ResUsers(models.Model):
 
     def write(self, vals):
         self.clear_caches()
+        print(vals)
         return super(ResUsers, self).write(vals)
 
     def _compute_check_user(self):
+        print("_compute_check_user")
         """To determine if the user has warehouse location restrictions.
         Sets the check_user field accordingly."""
         restriction_group_id = self.env.ref(